MATDAANSECURE_BACKEND/requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
mysql-connector-python==8.1.0
pyjwt==2.8.0
bcrypt==4.0.1
firebase-admin==6.1.0
flask-cors==3.0.10
gunicorn